Create a deployable storybook configuration.
The production image will contain a static storybook site that can be enabled by setting the environment variable NODE_ENV=storybook. The benefit of this approach is it prevents the need to have multiple docker images (production, storybook) and tightly binds the production image to the components available in the storybook. Downsides are slower builds and larger docker images.