Refactor to move away from handlebars view engine to a pure react solution.
Child components such as the ErrorPage can communicate with the context i.e. express server request to pass on information such as error codes. This allows components to be in control of the response however it does mean that the application must be rendered first to collect these data emitted from child components. React renderToNodeStream can only be used after the application has been statically rendered for SSR, consequently time to first byte will be delayed. Future updates to the react framework may introduce a suspend api which will allow for the necessary retrospective collecting of data behaviour and mean that the renderToStaticMarkup call can be eliminated.

To share data between two component trees on a per request basis, PageContext with a shared "bridge" class is used. This allows for manipulation of elements in the head and body. React cannot directly mount into the HEAD of the html document because external scripts my change HEAD child nodes outside of reacts control. For client side, further implementation needs to be done to support adding and deleteing nodes from the HEAD. As a proof of concept, the DOMBridge class providers a way to look at the children being set within the HEAD, identify and TITLE nodes and set the document title accordingly.